Project 1 M25/requirements.txt

flask
flask-cors
openai
python-dotenv
numpy
